Break down details.py into several files
